+++ launch/mission.launch
@@ -15,7 +15,7 @@
     <arg name="fileresult" default="1"/>
     <arg name="history_data" default="None"/>
     <arg name="history_decisions" default="None"/>
-    <arg name="task_scheduler" default="None"/>
+    <arg name="task_scheduler" default="central_heuristic_decision"/>
     <arg name="save" default="False"/>
 
     <param name="/use_sim_time" value="true" />
@@ -35,7 +35,7 @@
     </include>
 
     <include file="$(find int_preservation)/launch/task_scheduler.launch">
-        <arg name="method" value="$(arg task_scheduler)"/>
+        <arg name="task_scheduler" value="$(arg task_scheduler)"/>
     </include>
 
     <include file="$(find int_preservation)/launch/areas/area_sim$(arg nareas).launch">
