+++ launch/task_scheduler.launch
@@ -1,9 +1,9 @@
 <launch>
-    <arg name="task_scheduler" default="None"/>
+<arg name="task_scheduler"/>
 
-    <group unless="$(eval task_scheduler == 'None')">
-        <node name="$(arg task_scheduler)" pkg="int_preservation" type="$(arg task_scheduler).py">
-        </node>
-    </group>
+<group unless="$(eval task_scheduler == 'None')">
+    <node name="task_scheduler" pkg="int_preservation" type="$(arg task_scheduler).py">
+    </node>
+</group>
 
 </launch>
