+++ srv/clusterAssignment.srv
@@ -1,0 +1,3 @@
+int8[] cluster
+---
+bool availability
